assignment 1 pipeline's code
This is the completed code that was submitted for BCB420 Assignment 1 for the Bioinformatics and Computational biology specialization course. There is an associated HTML file that did knit from R and had errors when being shared but compiled locally. I will fix this on the next commits. Right now it does work but there will be improvements made to the efficiency, and appearance that bring it up to the standards that I have learned in my research courses.